backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
python-multipart==0.0.6
pytest==7.4.3
pytesseract==0.3.10
Pillow==10.1.0
ollama==0.1.33
aiofiles==23.2.1
httpx==0.25.2
